+++ ios/src/CountlyReactNative.h
@@ -44,6 +44,9 @@ typedef void (^Result)(id _Nullable result);
 - (void)getRemoteConfigValueForKey:(NSArray*_Nullable)arguments callback:(RCTResponseSenderBlock _Nullable)callback;
 - (void)showStarRating:(NSArray*_Nullable)arguments callback:(RCTResponseSenderBlock _Nullable)callback;
 - (void)showFeedbackPopup:(NSArray*_Nullable)arguments;
+- (void)presentFeedbackWidget:(NSArray*_Nullable)arguments;
+- (void)replaceAllAppKeysInQueueWithCurrentAppKey;
+- (void)removeDifferentAppKeysFromQueue;
 - (void)setEventSendThreshold:(NSArray*_Nullable)arguments;
 - (void)pushTokenType:(NSArray*_Nullable)arguments;
 - (void)sendPushToken:(NSArray*_Nullable)arguments;
@@ -64,4 +67,5 @@ typedef void (^Result)(id _Nullable result);
 
 - (void)enableAttribution;
 - (void)recordAttributionID:(NSArray*_Nullable)arguments;
+- (void)appLoadingFinished;
 @end